This Hangman Project
This a classic word guessing game , I have created it with Python Language. In this game we need to guess a correct letter in random word to save a man . If we guess a incorrect letter in a random word then you will lose a live . You will only have 6 Lives to crack the word . The game ends if we are out of Lives.