Project Explanation
This is a simple calculator project , I built using Python. It supports basic arithmetic operations like addition, subtraction, multiplication, and division. I designed it with a while loop so users can continue calculations using the previous result or start a new one whenever they want.
The project uses:
Functions for each math operation.
A dictionary to map operators to their functions.
A while loop to keep the calculator running interactively.
ASCII art (from a separate calculatorart.py file) to give it a fun, visual look.

I really enjoyed working on this it was both fun and a good challenge. It helped me improve my Python skills, especially with user input, loops, and function with inputs and outputs and also Dictionaries .